GUI based program in Java for for Eye to Hand Coordination
It is simple GUI based java program, built for the purpose to determine how effective your hands are coordinating with your eyes.
This program extends Java's JFrame class and also implements the MouseListener interface. When the program is executed it will show a circle of random color on a random position on the JFram panel. When you will click on the circle, it will disappear and another random colored circle at a random position will appear on the panel.
This circle continues 20 times and after that a window will pop up that will show your time in seconds to click 20 circles.