GUI based java program which utilize Action Listener
It is a simple GUI based java game, which extends JFrame class and implements Action Listener interface.
The GUI has a JFrame which serve as field for the ball in which it can move. Furthermore we have a JPanel in which we have four buttons through which you can move the ball in any of the four direction.